changing default_serial_number constant to bigInteger, addressing PR comments

removing unused method


changing default_serial_number constant to BigInteger and addressing PR comments

+++ hedera-mirror-web3/src/test/java/com/hedera/mirror/web3/service/AbstractContractCallServiceHistoricalTest.java
@@ -168,7 +168,7 @@ public abstract class AbstractContractCallServiceHistoricalTest extends Abstract
         domainBuilder
                 .nftHistory()
                 .customize(n -> n.tokenId(tokenEntity.getId())
-                        .serialNumber(DEFAULT_SERIAL_NUMBER)
+                        .serialNumber(DEFAULT_SERIAL_NUMBER.longValue())
                         .timestampRange(timestampRange))
                 .persist();
