+++ hedera-mirror-web3/src/test/java/com/hedera/mirror/web3/service/ContractCallNestedCallsTest.java
@@ -411,8 +411,8 @@ class ContractCallNestedCallsTest extends AbstractContractCallServiceOpcodeTrace
         final var contract = testWeb3jService.deploy(NestedCalls::deploy);
 
         // When
-        final var function = contract.call_nestedHtsGetApprovedAndHardcodedResult(
-                Address.ZERO.toHexString(), BigInteger.valueOf(DEFAULT_SERIAL_NUMBER));
+        final var function =
+                contract.call_nestedHtsGetApprovedAndHardcodedResult(Address.ZERO.toHexString(), DEFAULT_SERIAL_NUMBER);
         final var result = function.send();
 
         // Then
